add INT32_MAX to to_int()

+++ include/Book.hpp
@@ -36,6 +36,6 @@ public:
     BookName name;
     BookAuthor author;
     BookKeywords keywords;
-    int quantity;
+    ll quantity;
     double price;
 };
