+++ src/BigBrother.cpp
@@ -1,5 +1,6 @@
 #include "BigBrother.hpp"
 #include "Name.hpp"
+#include "CONSTEXPRS.hpp"
 namespace BigBrother
 {
     std::vector<Finance> tmp;
@@ -8,7 +9,7 @@ int BigBrother::show_finance(const std::vector<std::string> &S)
 {
     if (S.size() > 2)
         return -1;
-    int count;
+    ll count;
     if (S.size() == 1)
         count = cashLog().size();
     else if (NAME::to_int(S.at(1), count))
